Implemented Builder and Save methods for user and weightGoals

+++ src/entity/CommonUser.java
@@ -30,7 +30,7 @@ public class CommonUser implements User{
 
     private String paceType;
 
-    private int requiredCalories;
+    private double requiredCalories;
 
 
 
@@ -47,7 +47,7 @@ public class CommonUser implements User{
                       HashMap<String, Boolean> restrictions,
                       HashMap<String, Boolean> weightGoalType,
                       String paceType,
-                      int requiredCalories) {
+                      double requiredCalories) {
         this.userId = userId;
         this.name = name;
         this.password = password;
@@ -206,7 +206,7 @@ public class CommonUser implements User{
 
 
 
-    public int getRequiredCalories() {
+    public double getRequiredCalories() {
         return requiredCalories;
     }
 
@@ -261,7 +261,7 @@ public class CommonUser implements User{
     }
 
     @Override
-    public void setRequiredCalories(int requiredCalories) {
+    public void setRequiredCalories(double requiredCalories) {
         this.requiredCalories = requiredCalories;
     }
