+++ src/entity/User.java
@@ -49,7 +49,7 @@ public interface User {
 //
 //    HashMap<String, String> getPaceType();
 
-    int getRequiredCalories(); // Should default to 0
+    double getRequiredCalories(); // Should default to 0
 
     // setters
     void setPassword(String password);
@@ -73,6 +73,6 @@ public interface User {
     void setPaceType(String paceType);
 
 
-    void setRequiredCalories(int requiredCalories);
+    void setRequiredCalories(double requiredCalories);
 
 }
